+++ bindings/nostr-sdk-ffi/src/protocol/event/tag/mod.rs
@@ -151,6 +151,16 @@ impl Tag {
         }
     }
 
+    /// Construct `["client", "<name>"]` tag
+    ///
+    /// <https://github.com/nostr-protocol/nips/blob/master/89.md>
+    #[uniffi::constructor]
+    pub fn client(name: String) -> Self {
+        Self {
+            inner: tag::Tag::client(name),
+        }
+    }
+
     /// Compose `["expiration", "<timestamp>"]` tag
     ///
     /// <https://github.com/nostr-protocol/nips/blob/master/40.md>